requirements.txt.txt
langchain
langchain-core
langchain-huggingface
python-dotenv
transformers
torch
huggingface-hub